Hafta-5/backend/auth.py
from passlib.context import CryptContext
from jose import jwt
from datetime import datetime, timedelta

# Bu projede JWT demo amaçlıdır. Normalde üretim ortamında .env'den saklanır.
SECRET_KEY = "super_secret_super_safe_key_for_transit"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24 * 7 # 7 Günlük Token

# Parola şifreleme algoritması
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
