Hafta-5/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg[binary]
google-generativeai
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
